fix(get rooms): fixed api token room logic
There was an issue where api tokens with an unlimited duration where not being returned in the get rooms function.  This is because a token with an unlimited duration has an expiresAt  of Null, and the greater than query

now the query searches for either valid token or ones that have an unlimited duration (aka null).